requirements.txt
streamlit==1.39.0
semantic-kernel==1.36.0
openai>=1.35.0
PyPDF2==3.0.1
pdf2image==1.17.0
pillow==10.4.0
python-dotenv==1.0.1
markdown==3.6
